@@ -1,15 +1,18 @@
 <template>
-  <hello-world />
+  <div class="home">
+    <img alt="Vue logo" src="../assets/logo.png">
+    <HelloWorld msg="Welcome to Your Vue.js App"/>
+  </div>
 </template>
 
 <script>
-  import HelloWorld from '../components/HelloWorld'
+// @ is an alias to /src
+import HelloWorld from '@/components/HelloWorld.vue'
 
-  export default {
-    name: 'Home',
-
-    components: {
-      HelloWorld,
-    },
+export default {
+  name: 'HomeView',
+  components: {
+    HelloWorld
   }
+}
 </script>
