@@ -1,9 +1,4 @@
 const { defineConfig } = require('@vue/cli-service')
 module.exports = defineConfig({
-  transpileDependencies: [
-    'vuetify'
-  ],
-  devServer: {
-    proxy: "http://localhost:3000"
-  }
+  transpileDependencies: true
 })
